🌱 Split CloudProviderIcon: extract SVG icon components
Move the 13 inline provider SVG components (AWSIcon..KubernetesIcon)
and KIND_ICON_STYLE into CloudProviderIcon.icons.tsx. CloudProviderIcon.tsx
drops from 599 to 319 lines and now contains only the public API:
CloudProviderIcon, getProviderLabel, useProviderLabel, getProviderColor,
getProviderBorderClass, getConsoleUrl, detectCloudProvider.

Pure move — no markup, logic or exported-API change.
